requirements.txt
PyGithub==2.1.1
google-generativeai==0.3.1
python-dotenv==1.0.0
rich==13.7.0